+++ classDataPersonnel.php
@@ -1,66 +1,32 @@
-<?php
-
-    class Clients {
-        public $idClient;
-        private $nom;
-        private $telephone;
-        
-        
-        public $message;
-
-        function __construct($nom, $telephone) {
-            
-            $this->nom = $nom;
-            $this->telephone = $telephone;
-        }
-
-        
-        function insererClient() {
-            include 'connexion.php';
-            $sql = ("INSERT INTO DataPersonnel ( Nom, Telephone) values ('".$this->nom."', '".$this->telephone."')");
-            if(mysqli_query($db, $sql)){
-                //echo"<small style='color: green'>insertion fait</small>";
-                }else{
-                //echo "<small style='color: red;'>error : ".mysqli_error($db). " desolee</small>";
-                //return 'error'.mysqli_error($db);
-                $this->message = mysqli_error($db);
-            }
-        }
-        function updateClient() {
-            include 'connexion.php';
-
-            $updC1= ("UPDATE `DataPersonnel` SET `Nom` ='".$this->nom."' WHERE idDataPersonnel =$this->idClient");
-            if(mysqli_query($db,$updC1)){echo"";}else{
-                $this->message = mysqli_error($db);
-                return;
-            }
-            $updC2= ("UPDATE `DataPersonnel` SET `Telephone` = '".$this->telephone."' WHERE idDataPersonnel =$this->idClient");
-            if(mysqli_query($db,$updC2)){echo"";}else{
-                $this->message = mysqli_error($db);
-                return;
-            }
-        }
-
-        function deleteClient() {
-            include 'connexion.php';
-            $delete = ("DELETE FROM DataPersonnel WHERE idDataPersonnel =$this->idClient");
-            if (mysqli_query($db, $delete)){echo"";} else {
-                $this->message = mysqli_error($db);
-                return;
-            }
-        }
-
+<?php 
+    $user = "";
+    session_start();
+    if(isset($_GET['deconnexion']))
+    { 
+    if($_GET['deconnexion']==true)
+    {  
+        session_destroy();
+        header("location:index.php");
     }
+    }
+    else if($_SESSION['username'] !== ""){
+        $user = $_SESSION['username'];
+    }
+?>
 
+<?php
+    include 'write_read_json.php';
+    include 'true_classDataPersonnel.php';
     $q = $_REQUEST["q"];
     $tabC = explode("::", $q);
-    
+   
     $autre = '';
-    if (end($tabC) == 'add') {
+    if (end($tabC) == 'add' && $user != "") {
         if ($q !== "") {
             $hint = $q;
-            $tracteur = new Clients($tabC[0], $tabC[1]);
+            $tracteur = new Personnel($tabC[0], $tabC[1], $tabC[2], $tabC[3], $tabC[4]);
             $tracteur->insererClient();
+            $tracteur->write_insert();
             $autre = $tracteur->message;
             if( $tracteur->message) {
                 $hint = $autre;
@@ -77,12 +43,13 @@
     </div>';
         echo $hint == $autre ? $error : $sucess;
     } 
-    if (end($tabC) == 'update') {
+    if (end($tabC) == 'update' && $user != "") {
         if ($q !== "") {
             $hint = $q;
-            $tracteur = new Clients($tabC[0], $tabC[1]);
-            $tracteur->idClient = $tabC[2];
+            $tracteur = new Personnel($tabC[0], $tabC[1], $tabC[2], $tabC[3], $tabC[4]);
+            $tracteur->idClient = $tabC[5];
             $tracteur->updateClient();
+            $tracteur->write_update();
             $autre = $tracteur->message;
             if( $tracteur->message) {
                 $hint = $autre;
@@ -100,12 +67,13 @@
         echo $hint == $autre ? $error : $sucess;
     }
 
-    if (end($tabC) == 'delete') {
+    if (end($tabC) == 'delete' && $user != "") {
         if ($q !== "") {
             $hint = $q;
-            $tracteur = new Clients(0, 1);
+            $tracteur = new Personnel(0, 1, 2, 3, 4);
             $tracteur->idClient = $tabC[0];
             $tracteur->deleteClient();
+            $tracteur->write_delete();
             $autre = $tracteur->message;
             if( $tracteur->message) {
                 $hint = $autre;
